fix(Beamforming): folding multiple days does not break when a filterbank is not written due to low data fraction
When there is less than 50% data fraction during beamforming, the skybeam object is saved as None. When you try to create a filterbank from the skybeam, it fails, which stops the rest of the observations from getting folded. Now, observations with less than 50% data fraction are still not folded on, but they are skipped gracefully instead of exiting the loop.