104 multi day search stops when data fraction 05 (#106)
* fix(multiday_search): sets minimum number of f0 bins in multiday search to 2

* fix(multiday_search): prevents logs from being grabbed when choosing raw data to beam form

* fix(Beamforming): folding multiple days does not break when a filterbank is not written due to low data fraction

When there is less than 50% data fraction during beamforming, the skybeam object is saved as None. When you try to create a filterbank from the skybeam, it fails, which stops the rest of the observations from getting folded. Now, observations with less than 50% data fraction are still not folded on, but they are skipped gracefully instead of exiting the loop.

* feat(load_profiles): small change to how archives are printed

* fix(confirm_cand): limits number of observations searched to number of days folded

* Add Dockr Swarm Service task inspection to end of log output

* feat(multidayfold_pipeline): default number of days folded and searched is now all that are available

To fold a specific number of days, use flag 'nday'.

* feat(fold_multiday): default option for nday is zero

This is a cleaner scheme, the default is still to fold all days but now it will not be redundant in the case of >=999 days to fold :)

* fix(confirm_cand): now fails gracefully if not enough archives to search through

---------

Co-authored-by: Magnus L'Argent <magnuslargent@sps-archiver1.chime>
Co-authored-by: chrisfandrade16 <chrisfandrade16@gmail.com>